requirements.txt
pandas==1.5.2
undetected-chromedriver==3.1.7
spacy==3.0.6
Unidecode==1.3.6
pytest==7.3.1
PyYAML==6.0